requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
openai-whisper>=20230314
torch>=2.0
ffmpeg-python>=0.2
python-multipart>=0.0.7

